requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
aiosqlite
apscheduler
python-dotenv
lxml
